lession 9 of upskilling in Next.js
